services/user-mgmt-service/requirements.txt
django
djangorestframework
mysqlclient
djangorestframework-simplejwt
cryptography